requirements_cpu_1.txt
-f https://download.pytorch.org/whl/cpu
cached_property
numpy_groupies
opencv_python
pandas
pytest
PyYAML
scikit-learn==1.0.2
scikit-image==0.19.3
scipy==1.8.0
tqdm
numpy==1.22.4
torch==2.0.1
torchvision==0.15.2